+++ NEWS.md
@@ -34,6 +34,7 @@ libFirm 1.22.1 (2016-01-07)
 * ir: Slightly improve preservation of debug info during transformations
 * amd64: Improve use of `lea` in instruction selection
 * sparc: Support the asm constaint `K`
+* sparc: Support computed goto
 * Bugfixes
 
 libFirm 1.22.0 (2015-12-31)